requirements.txt
fastapi
uvicorn
motor
pydantic
python-dotenv
fastapi-mqtt
gmqtt